Final: Added Music, Updated Window and Stage Classes and Diagramas Class
This is the final update before turning the assignment in.
Added music to the game, although I wasn't able to make a function to stop playing it.
Updated Window and Stage Classes
Updated Diagram Class
IMPORTANT: Please, read the instructions within the SQL functions in the Stage Class, for the Schema is necessary to get the DB working properly.